add NotFoundException for endpoint with id

+++ src/products/products.controller.ts
@@ -1,5 +1,5 @@
 /* eslint-disable prettier/prettier */
-import { Controller, Get, Param, Delete, Body, Post, Put } from '@nestjs/common';
+import { Controller, Get, Param, Delete, Body, Post, Put, NotFoundException } from '@nestjs/common';
 import { ProductsService } from './products.service';
 import { CreateProductDTO } from './dtos/create-product.dto';
 import { ParseUUIDPipe } from '@nestjs/common';
@@ -17,11 +17,17 @@ export class ProductsController {
 	@Get('/:id')
 	//public getById(@Param('id') id: string) {
     getById(@Param('id', new ParseUUIDPipe()) id: string) {
+      if (!this.productsService.getById(id))
+      throw new NotFoundException('Product not found');
+
 		return this.productsService.getById(id);
 	};
 
 	@Delete('/:id')
   deleteById(@Param('id', new ParseUUIDPipe()) id: string) {
+    if (!this.productsService.getById(id))
+      throw new NotFoundException('Product not found');
+
     this.productsService.deleteById(id);
     return { success: true };
   };
@@ -36,6 +42,9 @@ export class ProductsController {
     @Param('id', new ParseUUIDPipe()) id: string,
     @Body() productData: UpdateProductDTO,
   ) {
+    if (!this.productsService.getById(id))
+      throw new NotFoundException('Product not found');
+
     this.productsService.updateById(id, productData);
     return { success: true };
   }
